refactor: extract crud dialog overlay into separate file
Move the internal CrudDialogOverlay element out of vaadin-crud-dialog.js
into its own vaadin-crud-dialog-overlay.js. No behavior change.
